+++ src/main/java/com/example/mfsp/service/orderService.java
@@ -5,9 +5,9 @@ import com.example.mfsp.entity.Orderform;
 import java.util.List;
 
 public interface orderService {
-    public long addOrder(Orderform orderform);
-    public long deleteOrder(Orderform orderform);
+    public int addOrder(Orderform orderform);
+    public int deleteOrder(Orderform orderform);
     public List<Orderform> selAllOrder();
-    public long updateOrder(Orderform orderform);
+    public int updateOrder(Orderform orderform);
 
 }
